Move plugin modules
