+++ src/amulet/app/plugin/_container.py
