+++ src/amulet/app/plugin/_data.py
