+++ src/amulet/app/plugin/_plugin.py
