+++ src/amulet/app/plugin/_state.py
