+++ src_/builtin_plugins_/amulet_team_main_window/_plugin.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from amulet_team_main_window.application.windows.main_window import AmuletMainWindow
-from amulet_editor.models.plugin import PluginV1
+from amulet.app.plugin import PluginV1
 
 
 window: AmuletMainWindow | None = None
