+++ src_/builtin_plugins_/amulet_team_settings/_plugin.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from PySide6.QtCore import Qt
 
-from amulet_editor.models.plugin import PluginV1
+from amulet.app.plugin import PluginV1
 import tablericons
 
 from .settings import SettingsPage
